requirements.txt
joblib==1.4.2
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
scikit-learn==1.5.1
seaborn==0.13.2
streamlit==1.38.0